# Faster unsigned division by constants