… THIRD-PARTY-NOTICES.TXT
… src/coreclr/jit/lsraxarch.cpp
… src/coreclr/jit/utils.cpp
… src/coreclr/jit/utils.h
… src/tests/JIT/Math/Functions/DivideByConst.cs
… src/tests/JIT/Math/Functions/Program.cs