Improved magic number calculation based on "Faster Unsigned Division by Constants" by ridiculous_fish.
https://ridiculousfish.com/files/faster_unsigned_division_by_constants.pdf
https://github.com/ridiculousfish/libdivide/blob/master/doc/divide_by_constants_codegen_reference.c

Raw performance is 12-40% faster for "uncooperative" divisors and register pressure is also reduced.
`DateTime.Year` for example is 38% faster.

Perf stats on a i7-7700K CPU 4.20GHz:

| Method |     Main |     PR |    Speedup |
|------- |---------:|----------:|----------:|
|     3 | 1.129 ns | 1.113 ns | 1.01 |
|     7 | 1.792 ns | 1.113 ns | 1.61 |
|    10 | 1.129 ns | 1.112 ns | 1.02 |
|    14 | 1.790 ns | 1.113 ns | 1.61 |
|    56 | 1.791 ns | 1.112 ns | 1.61 |
|     3ul | 1.112 ns | 1.112 ns | |
|     7ul | 1.798 ns | 1.595 ns | 1.13 |
|    10ul | 1.112 ns | 1.112 ns | |
|    14ul | 1.799 ns | 1.336 ns | 1.35 |
|    56ul | 1.800 ns | 1.112 ns | 1.62 |
| DateTime.Year | 5.756 ns | 3.599 ns | 1.60 |