@tannergooding 2021-03-13
Still a draft so I'm not going to review in depth yet. I'd definitely be interested in seeing some perf numbers and assembly diffs 😄

I'd also be interested in how this compares to say `udiv.exe` and `sdiv.exe` which AMD provides code listings for in their software optimization manual (https://www.amd.com/system/files/TechDocs/47414_15h_sw_opt_guide.pdf, 9.6 Derivation of Algorithm, Multiplier, and Shift 
Factor for Integer Division by Constants). These programs are themselves based on `Granlund, T.; Montgomery,  P.L.: "Division by Invariant Integers using Multiplication."  SIGPLAN Notices, Vol. 29, June 1994, page 61.` and ` Magenheimer, D.J.; et al: "Integer Multiplication and Division on the HP Precision Architecture." IEEE Transactions on Computers, Vol. 37, No. 8, August 1988, page 980.`