@GrabYourPitchforks 2021-03-13
Somebody smarter than me needs to validate license information. There's license text at the top of https://github.com/ridiculousfish/libdivide/blob/master/doc/divide_by_constants_codegen_reference.c which does not match the license text within https://github.com/ridiculousfish/libdivide/blob/master/LICENSE.txt.