@pentp 2021-03-13
That should be the current implementation from https://github.com/dotnet/coreclr/pull/10996 (at least the comments in code referenced the same paper by Torbjorn Granlund and Peter L. Montgomery).
There are some microbenchmark perf numbers in the linked PDF by ridiculous_fish (10-15% faster for "uncooperative" divisors), but in real code the gains could be even bigger because of reduced register pressure.

This is the first time I'm dealing with ARM assembly and I'm having trouble figuring out how to emit conditional `mvnhs` (clang does it somehow). Do I need to extend `emitIns_R_R_COND` to add support for this in Arm64 and create an entirely new emit method for `A32`?